Crate zen_mesh

This crate can deserialize .mrm or .msh meshes, and convert them into gltf files.

use std::{convert::TryFrom, fs::File, io::Cursor};
use zen_archive::Vdfs;
use zen_mesh::{gltf, mrm::MrmMesh, GeneralMesh};
use zen_types::path::INSTANCE;

let vdf_file = File::open(INSTANCE.meshes())?;
let vdf = Vdfs::new(vdf_file)?;
let mesh_entry = vdf
    .get_by_name("ORC_MASTERTHRONE.MRM")
    .expect("Should be there!");
let cursor = Cursor::new(mesh_entry.data);
let mesh = MrmMesh::new(cursor, "ORC_MASTERTHRONE")?;
let mesh = GeneralMesh::try_from(mesh)?;
let _gltf = gltf::to_gltf(mesh, gltf::Output::Binary);

Re-exports

pub use mrm::MrmMesh;

pub use msh::MshMesh;

Modules

gltf

mrm

msh

structures

Structs

GeneralMesh
General Mesh consisting out of one or more sub meshes

Mesh
Basic Mesh Informations

SubMesh
Mesh that is a component in another mesh, already holds its material

Enums

Error